Cleaned up DrawState stuff into vsvg-cli
